refactor(parser): address review — own module, all same-named sets, docs
Three review findings on #1968:

- Greptile P1 / Codex P2: an occurrence may carry several property sets of one
  name (one per IfcRelDefinesByProperties; nothing dedupes them). The merge
  indexed only the first, so a twin set stayed bare — and IDS requires EVERY set
  matching the facet's propertySet constraint to satisfy it, which reintroduces
  the same false "not found" this rule exists to prevent. Inherited properties
  now merge into every same-named set.
- Codex P1: the helper moves to its own `property-set-merge.ts` rather than
  growing an already 1167-line module past the ~400-line rule. The rule is
  independent of how either side was extracted, so it reads better alone.
- Codex P1: document the new public export. `docs/guide/parsing.md` gains a
  "Type-Inherited Properties" section stating the per-property precedence, the
  duplicate-set behaviour and the immutability contract. The snippet is
  typechecked by `pnpm docs:check-samples`.
