+++ packages/parser/src/on-demand-extractors.ts
@@ -462,64 +462,6 @@ export function extractTypePropertiesOnDemand(
     };
 }
 
-/**
- * Minimal shape the type/occurrence property merge needs. `properties` is
- * `readonly` so consumers whose own type declares it as a `ReadonlyArray`
- * (`@ifc-lite/lens`'s `PropertySetInfo`) still satisfy the constraint and
- * infer their own element type rather than falling back to this one.
- */
-interface NamedPropertySet {
-    name: string;
-    properties: readonly { name: string }[];
-}
-
-/**
- * Compose an occurrence's own property sets with those it inherits from its
- * `IfcTypeProduct`, the way IFC defines the inheritance: **per property, not
- * per property set**.
- *
- * An occurrence and its type routinely both carry a set of the same name —
- * `Pset_CoveringCommon` holding `IsExternal`/`Reference` on the occurrence and
- * `SurfaceSpreadOfFlame`/`Combustible` on the type is a plain Revit export.
- * Dropping the whole inherited set on a name collision makes every type-only
- * property in it invisible, which is what made IDS report a present property as
- * missing (#1913). Where both define the same property name, the occurrence
- * wins — it is the more specific definition.
- *
- * Neither input is mutated: a set that gains inherited properties is returned
- * as a new object, so cached extractor results stay intact.
- */
-export function mergeInheritedPropertySets<T extends NamedPropertySet>(
-    ownSets: readonly T[],
-    inheritedSets: readonly T[],
-): T[] {
-    if (inheritedSets.length === 0) return ownSets.slice();
-
-    const merged = ownSets.slice();
-    const byName = new Map<string, number>();
-    for (let i = 0; i < merged.length; i++) {
-        if (!byName.has(merged[i].name)) byName.set(merged[i].name, i);
-    }
-
-    for (const inherited of inheritedSets) {
-        const index = byName.get(inherited.name);
-        if (index === undefined) {
-            merged.push(inherited);
-            byName.set(inherited.name, merged.length - 1);
-            continue;
-        }
-        const own = merged[index];
-        const ownPropNames = new Set(own.properties.map((p) => p.name));
-        const additions = inherited.properties.filter((p) => !ownPropNames.has(p.name));
-        if (additions.length === 0) continue;
-        // `as T`: the spread reproduces every field of `own` and replaces only
-        // `properties` with a superset of the same element type, but TypeScript
-        // cannot narrow a spread of an unresolved generic back to that generic.
-        merged[index] = { ...own, properties: [...own.properties, ...additions] } as T;
-    }
-
-    return merged;
-}
 
 /**
  * Extract properties from a type entity's own HasPropertySets attribute.
